pyproject.toml
[project]
name = "health-care-development"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.2.0",
    "cloudinary>=1.44.1",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cloudinary-storage>=0.3.0",
    "django-encrypted-model-fields>=0.6.5",
    "django-model-utils>=5.0.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "daphne>=4.2.1",
    "django-debug-toolbar>=6.3.0",
    "whitenoise>=6.8.2",
    "django-auditlog>=3.4.1",
]
